1.修改：OSAL的cmake, 改为 osal/FileLists.cmake 文件, 在 linux_simulator/CMakeLists.txt 中通过包含 osal/FileLists.cmake 添加 OSAL 的源文件。

+++ osal/core/osal_mem.c
@@ -191,10 +191,10 @@ void OsalMemInit(uint32_t * addr, size_t size_bytes)
 	
 	addr_tmp = (uint8_t *)addr;
 	
-	if(((uint32_t)addr) & (HEAP_ALIGNMENT_SIZE - 1))
+	if(((size_t)addr) & (HEAP_ALIGNMENT_SIZE - 1))
 	{
-		addr_tmp += HEAP_ALIGNMENT_SIZE - (((uint32_t)addr) & (HEAP_ALIGNMENT_SIZE - 1));
-		size_bytes -= (((uint32_t)addr) & (HEAP_ALIGNMENT_SIZE - 1));
+		addr_tmp += HEAP_ALIGNMENT_SIZE - (((size_t)addr) & (HEAP_ALIGNMENT_SIZE - 1));
+		size_bytes -= (((size_t)addr) & (HEAP_ALIGNMENT_SIZE - 1));
 	}
 
 	g_pMemHead = (osal_mem_head_t *)addr_tmp;
